Initial implementation of the library
This is the first commit of the simple and direct implementation of
the work pool. The aim of the library is not to save resources by
reusing goroutines, it is to limit the resources that can pottentially
be used while creating goroutines without any control.

Check the README for more information.